pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "wednesday-frog"
version = "0.1.0"
description = "Self-hosted frog image broadcaster for Slack, Teams, Mattermost, Discord, and Zoom."
readme = "README.md"
requires-python = ">=3.12"
license = { text = "Apache-2.0" }
authors = [{ name = "Hero of Time and Space" }]
dependencies = [
  "alembic>=1.14,<2.0",
  "apscheduler>=3.10,<4.0",
  "argon2-cffi>=23.1,<24.0",
  "cryptography>=44.0,<45.0",
  "fastapi>=0.115,<1.0",
  "httpx>=0.28,<1.0",
  "itsdangerous>=2.2,<3.0",
  "jinja2>=3.1,<4.0",
  "pillow>=11.0,<12.0",
  "psycopg[binary]>=3.2,<4.0",
  "python-multipart>=0.0.9,<0.1",
  "sqlalchemy>=2.0,<3.0",
  "uvicorn[standard]>=0.34,<1.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3,<9.0",
]

[project.scripts]
wednesday-frog = "wednesday_frog.__main__:main"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
wednesday_frog = ["py.typed", "plugins/*/manifest.json"]

[tool.pytest.ini_options]
testpaths = ["tests"]
